@@ -1,9 +1,9 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext {
-        compose_version = '1.0.1'
+        compose_version = '1.1.1'
+        kotlin_version = '1.6.10'
     }
-    ext.kotlin_version = "1.4.31"
     ext.objectboxVersion = '2.7.1'
     repositories {
         google()
@@ -12,10 +12,12 @@ buildscript {
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.2"
-        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath("androidx.navigation:navigation-safe-args-gradle-plugin:2.3.5")
         classpath "io.objectbox:objectbox-gradle-plugin:$objectboxVersion"
         classpath "org.jacoco:org.jacoco.core:0.8.6"
+        def nav_version = "2.4.1"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
