@@ -5,20 +5,18 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.jeanloth.project.android.kotlin.axounaut.R
-import com.jeanloth.project.android.kotlin.domain_models.entities.ArticleWrapper
-import kotlinx.android.synthetic.main.item_checkbox_article.view.*
+import com.jeanloth.project.android.kotlin.axounaut.databinding.ItemCheckboxArticleBinding
 import com.jeanloth.project.android.kotlin.domain_models.entities.IngredientWrapper
 
-
 class CheckboxTextViewAdapter(
-    private var items : MutableList<IngredientWrapper>
-) : RecyclerView.Adapter<CheckboxTextViewAdapter.ItemHolder>()  {
+    private var items: MutableList<IngredientWrapper>
+) : RecyclerView.Adapter<CheckboxTextViewAdapter.ItemHolder>() {
 
-    var onCheckedItem : ((IngredientWrapper, Boolean) -> Unit)? = null
+    var onCheckedItem: ((IngredientWrapper, Boolean) -> Unit)? = null
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemHolder {
         val itemView = LayoutInflater.from(parent.context)
-                .inflate(R.layout.item_checkbox_article, parent, false)
+            .inflate(R.layout.item_checkbox_article, parent, false)
         return ItemHolder(itemView)
     }
 
@@ -31,28 +29,30 @@ class CheckboxTextViewAdapter(
         return items.size
     }
 
-    fun setItems(items : List<IngredientWrapper>){
+    fun setItems(items: List<IngredientWrapper>) {
         this.items = items.toMutableList()
         notifyDataSetChanged()
     }
 
     inner class ItemHolder(view: View) : RecyclerView.ViewHolder(view) {
+        private val binding = ItemCheckboxArticleBinding.bind(view)
+        fun bind(item: IngredientWrapper) {
 
-        fun bind(item : IngredientWrapper){
+            val articleLabel =
+                itemView.context.getString(R.string.article_name, item.ingredient.label)
 
-            val articleLabel = itemView.context.getString(R.string.article_name, item.ingredient.label)
-
-            itemView.tv_label.text= articleLabel
-
-            itemView.tv_label.setOnClickListener {
-                if(itemView.cb_item.isEnabled)
-                    itemView.cb_item.isChecked = !itemView.cb_item.isChecked
+            binding.tvLabel.apply {
+                text = articleLabel
+                setOnClickListener {
+                    if (binding.cbItem.isEnabled)
+                        binding.cbItem.isChecked = !binding.cbItem.isChecked
+                }
             }
 
-            itemView.cb_item.setOnCheckedChangeListener{ _, isChecked ->
+            binding.cbItem.setOnCheckedChangeListener { _, isChecked ->
                 onCheckedItem?.invoke(item, isChecked)
             }
         }
-        }
     }
+}
 
