@@ -9,10 +9,9 @@ import androidx.appcompat.content.res.AppCompatResources.getDrawable
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
 import com.jeanloth.project.android.kotlin.axounaut.R
+import com.jeanloth.project.android.kotlin.axounaut.databinding.ItemArticleListBinding
 import com.jeanloth.project.android.kotlin.domain_models.entities.Article
 import com.jeanloth.project.android.kotlin.domain_models.entities.ArticleWrapper
-import kotlinx.android.synthetic.main.item_article.view.tv_name
-import kotlinx.android.synthetic.main.item_article_list.view.*
 import splitties.views.onClick
 
 class ArticleListAdapter(
@@ -46,18 +45,20 @@ class ArticleListAdapter(
 
     inner class ArticleHolder(view: View) : RecyclerView.ViewHolder(view), View.OnClickListener {
 
+        private val binding = ItemArticleListBinding.bind(view)
+
         fun bind(article : Article){
-            itemView.tv_name.text= article.label
-            itemView.tv_price.text= context.getString(R.string.price_euro, article.price.toString())
+            binding.tvName.text= article.label
+            binding.tvPrice.text= context.getString(R.string.price_euro, article.price.toString())
             //itemView.tv_article_stat.text= context.getString(R.string.article_command_number, article.price.toString())
 
-            itemView.bt_edit_article.setOnClickListener {
+            binding.btEditArticle.setOnClickListener {
                 onEditClick?.invoke(article)
             }
 
-            itemView.bt_expand.onClick {
-                itemView.gp_article_details.visibility = if(itemView.gp_article_details.isVisible) View.GONE else View.VISIBLE
-                itemView.bt_expand.background = getDrawable(context, if(itemView.gp_article_details.isVisible) R.drawable.ic_up_arrow else R.drawable.ic_down_arrow)
+            binding.btExpand.onClick {
+                binding.gpArticleDetails.visibility = if(binding.gpArticleDetails.isVisible) View.GONE else View.VISIBLE
+                binding.btExpand.background = getDrawable(context, if(binding.gpArticleDetails.isVisible) R.drawable.ic_up_arrow else R.drawable.ic_down_arrow)
             }
         }
 
