@@ -2,7 +2,51 @@ package com.jeanloth.project.android.kotlin.axounaut.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Purple700 = Color(0xFF3700B3)
-val Teal200 = Color(0xFF03DAC5)
+val purple_200 = Color(0xFFBB86FC)
+val purple_500 = Color(0xFF6200EE)
+val purple_700 = Color(0xFF3700B3)
+val teal_200 = Color(0xFF03DAC5)
+val teal_700 = Color(0xFF018786)
+val black = Color(0xFF000000)
+val white = Color(0xFFFFFFFF)
+
+val red_001 = Color(0xFFDC0C3D)
+val red_002 = Color(0xFFAEBC2222)
+val salamander = Color(0xFFF05E23)
+val salamander_attenue = Color(0xFFF17746)
+val ginger = Color(0xFFBE5504)
+val sandstone = Color(0xFFD7722C)
+val orange_light = Color(0xFF70DD9D1D)
+val orange_001 = Color(0xFFE1A735)
+val orange_002 = Color(0xFFC6932E)
+val orange_003 = Color(0xFFB5862B)
+val orange_004 = Color(0xFFEAB64F)
+val flax = Color(0xFFDDCA7D)
+val orange_005 = Color(0xFFE1A735)
+
+val marron_dark_1 = Color(0xFF625337)
+val marron_light_1 = Color(0xFF91AB8D53)
+val marron_golden_1 = Color(0xFFA27035)
+
+val see_green = Color(0xFF2E8B57)
+val kelly_green = Color(0xFF4CBB17)
+val kelly_green_002 = Color(0xFF29BF12)
+val apple_green = Color(0xFF8DB500)
+val green_light_1 = Color(0xFF6D906A)
+val green_light_2 = Color(0xFF5A7758)
+val green_dark_1 = Color(0xFF0E5E20)
+val green_dark_2 = Color(0xFF1C5C2A)
+
+val gray_light = Color(0xFFE9EAE7)
+val gray_0 = Color(0xFFEDEFE9)
+val gray_1 = Color(0xFFB3B5AE)
+val gray_2 = Color(0xFF747571)
+val brown_57 = Color(0xFF91625337)
+
+// Other colors
+val blue_001 = Color(0xFF08BDBD)
+val light_blue_600 = Color(0xFF039BE5)
+val light_blue_900 = Color(0xFF01579B)
+val light_blue_A200 = Color(0xFF40C4FF)
+val light_blue_A400 = Color(0xFF00B0FF)
+val black_overlay = Color(0x66000000)
