@@ -7,18 +7,13 @@ import android.view.View
 import android.view.View.GONE
 import android.view.View.VISIBLE
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import com.jeanloth.project.android.kotlin.axounaut.R
 import com.jeanloth.project.android.kotlin.axounaut.adapters.ArticleListAdapter
-import com.jeanloth.project.android.kotlin.axounaut.extensions.openPopUpMenu
+import com.jeanloth.project.android.kotlin.axounaut.databinding.FragmentArticleBinding
 import com.jeanloth.project.android.kotlin.axounaut.viewModels.ArticleVM
 import com.jeanloth.project.android.kotlin.axounaut.viewModels.MainVM
-import kotlinx.android.synthetic.main.fragment_article.*
-import kotlinx.android.synthetic.main.item_article_list.*
-import kotlinx.android.synthetic.main.layout_header.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
 
@@ -34,6 +29,8 @@ class ArticleFragment : Fragment() {
     private val articleVM : ArticleVM by viewModel()
     private lateinit var adapter: ArticleListAdapter
 
+    private lateinit var binding: FragmentArticleBinding
+
     enum class EArticleDisplayMode{
         EDITION,
         DETAILLED
@@ -43,8 +40,8 @@ class ArticleFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_article, container, false)
+        binding = FragmentArticleBinding.inflate(layoutInflater, container, false)
+        return binding.root
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -54,9 +51,9 @@ class ArticleFragment : Fragment() {
         adapter = ArticleListAdapter(emptyList(), requireContext()).apply {
             onEditClick = { }
         }
-        rv_articles_fragment.adapter = adapter
+        binding.rvArticlesFragment.adapter = adapter
 
-        cl_add_article_btn.setOnClickListener{
+        binding.clAddArticleBtn.root.setOnClickListener{
             goToArticleDetails()
         }
 
@@ -64,7 +61,7 @@ class ArticleFragment : Fragment() {
             articleVM.allArticlesLiveData().observe(viewLifecycleOwner){
                 Log.d("[Article Fragment", "Article observed : $it")
                 adapter.setItems(it)
-                tv_error_no_articles.visibility = if(it.isEmpty()) VISIBLE else GONE
+                binding.tvErrorNoArticles.visibility = if(it.isEmpty()) VISIBLE else GONE
             }
         }
 
