@@ -8,16 +8,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.snackbar.Snackbar
-import com.jeanloth.project.android.kotlin.axounaut.MainActivity
-import com.jeanloth.project.android.kotlin.axounaut.R
+import com.jeanloth.project.android.kotlin.axounaut.databinding.FragmentClientDetailsBinding
 import com.jeanloth.project.android.kotlin.axounaut.extensions.isPhoneValid
 import com.jeanloth.project.android.kotlin.axounaut.viewModels.ClientVM
 import com.jeanloth.project.android.kotlin.axounaut.viewModels.MainVM
 import com.jeanloth.project.android.kotlin.domain_models.entities.AppClient
-import kotlinx.android.synthetic.main.fragment_client_details.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
-
 
 /**
  * A simple [Fragment] subclass.
@@ -28,12 +25,15 @@ class ClientDetailsFragment : Fragment() {
     private val mainVM : MainVM by sharedViewModel()
     private val clientVM : ClientVM by viewModel()
 
+    private lateinit var binding: FragmentClientDetailsBinding
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_client_details, container, false)
+        binding = FragmentClientDetailsBinding.inflate(layoutInflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -41,16 +41,16 @@ class ClientDetailsFragment : Fragment() {
 
         mainVM.setHeaderTitle("Ajouter un client")
 
-        bt_add_client.setOnClickListener {
+        binding.btAddClient.setOnClickListener {
             addClient()
         }
     }
 
     private fun addClient() {
-        val clientName = et_client_firstname.text.toString()
-        val clientLastname = et_client_lastname.text.toString()
-        val clientPhoneNumber = et_client_phone_number.text.toString()
-        val clientFidelity = et_client_fidelity.text.toString()
+        val clientName = binding.etClientFirstname.text.toString()
+        val clientLastname = binding.etClientLastname.text.toString()
+        val clientPhoneNumber = binding.etClientPhoneNumber.text.toString()
+        val clientFidelity = binding.etClientFidelity.text.toString()
 
         if( clientName.isEmpty() || !clientPhoneNumber.isPhoneValid()) {
             Snackbar.make(requireView(), "Veuillez saisir des valeurs valides.",
